Move all private methods from collections controller to items helper
